# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ParkingPaTi API
  version: 1.0.0
  description: Contrato base de la API REST de ParkingPaTi.
paths:
  /api/v1/admin/accounts/:
    get:
      operationId: admin_accounts_list
      description: |-
        Los controladores de este proyecto son viewsets.ViewSet "planos", no
        GenericViewSet - por eso DEFAULT_PAGINATION_CLASS (settings) NO se
        aplica automaticamente a sus list(). Este mixin da una forma de una
        sola linea para paginar manualmente en los list() que devuelven
        datasets que crecen con el tiempo (hallazgo 4.7 de la auditoria).

        Uso:
            class MiViewSet(PaginacionManualMixin, viewsets.ViewSet):
                def list(self, request):
                    queryset = MiRepository.listar()
                    return self.paginar(request, queryset, MiDTO)
      parameters:
      - in: query
        name: activo
        schema:
          type: boolean
      - in: query
        name: onboarding_estado
        schema:
          enum:
          - CORREO_PENDIENTE
          - DATOS_INICIALES_PENDIENTES
          - REVISION_PENDIENTE
          - RECHAZADO
          - CONFIGURACION_PENDIENTE
          - ACTIVO
          - DESHABILITADO
          type: string
          minLength: 1
        description: |-
          * `CORREO_PENDIENTE` - Correo pendiente
          * `DATOS_INICIALES_PENDIENTES` - Datos iniciales pendientes
          * `REVISION_PENDIENTE` - Revision pendiente
          * `RECHAZADO` - Rechazado
          * `CONFIGURACION_PENDIENTE` - Configuracion pendiente
          * `ACTIVO` - Activo
          * `DESHABILITADO` - Deshabilitado
      - in: query
        name: q
        schema:
          type: string
          maxLength: 150
          minLength: 1
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCuentaPaginaDTO'
          description: ''
  /api/v1/admin/accounts/{cuenta_id}/:
    get:
      operationId: admin_accounts_retrieve
      parameters:
      - in: path
        name: cuenta_id
        schema:
          type: integer
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCuentaDetalleDTO'
          description: ''
  /api/v1/admin/accounts/{cuenta_id}/disable/:
    post:
      operationId: admin_accounts_disable
      parameters:
      - in: path
        name: cuenta_id
        schema:
          type: integer
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccionResponseDTO'
          description: ''
  /api/v1/admin/applications/:
    get:
      operationId: admin_applications_list
      description: |-
        Los controladores de este proyecto son viewsets.ViewSet "planos", no
        GenericViewSet - por eso DEFAULT_PAGINATION_CLASS (settings) NO se
        aplica automaticamente a sus list(). Este mixin da una forma de una
        sola linea para paginar manualmente en los list() que devuelven
        datasets que crecen con el tiempo (hallazgo 4.7 de la auditoria).

        Uso:
            class MiViewSet(PaginacionManualMixin, viewsets.ViewSet):
                def list(self, request):
                    queryset = MiRepository.listar()
                    return self.paginar(request, queryset, MiDTO)
      parameters:
      - in: query
        name: onboarding_estado
        schema:
          enum:
          - CORREO_PENDIENTE
          - DATOS_INICIALES_PENDIENTES
          - REVISION_PENDIENTE
          - RECHAZADO
          - CONFIGURACION_PENDIENTE
          - ACTIVO
          - DESHABILITADO
          type: string
          minLength: 1
        description: |-
          * `CORREO_PENDIENTE` - Correo pendiente
          * `DATOS_INICIALES_PENDIENTES` - Datos iniciales pendientes
          * `REVISION_PENDIENTE` - Revision pendiente
          * `RECHAZADO` - Rechazado
          * `CONFIGURACION_PENDIENTE` - Configuracion pendiente
          * `ACTIVO` - Activo
          * `DESHABILITADO` - Deshabilitado
      - in: query
        name: q
        schema:
          type: string
          maxLength: 150
          minLength: 1
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSolicitudPaginaDTO'
          description: ''
  /api/v1/admin/applications/{cuenta_id}/:
    get:
      operationId: admin_applications_retrieve
      parameters:
      - in: path
        name: cuenta_id
        schema:
          type: integer
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSolicitudDetalleDTO'
          description: ''
  /api/v1/admin/applications/{cuenta_id}/approve/:
    post:
      operationId: admin_applications_approve
      parameters:
      - in: path
        name: cuenta_id
        schema:
          type: integer
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccionResponseDTO'
          description: ''
  /api/v1/admin/applications/{cuenta_id}/document/:
    get:
      operationId: admin_applications_document_retrieve
      parameters:
      - in: path
        name: cuenta_id
        schema:
          type: integer
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDocumentoAccesoDTO'
          description: ''
  /api/v1/admin/applications/{cuenta_id}/reject/:
    post:
      operationId: admin_applications_reject
      parameters:
      - in: path
        name: cuenta_id
        schema:
          type: integer
        required: true
      tags:
      - admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRechazoDTO'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminRechazoDTO'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminRechazoDTO'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccionResponseDTO'
          description: ''
...
components:
  schemas:
    AdminAccionResponseDTO:
      type: object
      properties:
        detail:
          type: string
        cuenta_id:
          type: integer
        onboarding_estado:
          type: string
        email_enviado:
          type: boolean
      required:
      - cuenta_id
      - detail
      - onboarding_estado
...
    AdminCuentaDetalleDTO:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        correo:
          type: string
          format: email
          maxLength: 254
        persona:
          $ref: '#/components/schemas/AdminPersonaDTO'
        is_active:
          type: boolean
          title: Activo
          description: Indica si el usuario debe ser tratado como activo. Desmarque
            esta opción en lugar de borrar la cuenta.
        correo_verificado:
          type: boolean
        onboarding_estado:
          $ref: '#/components/schemas/OnboardingEstadoEnum'
        date_joined:
          type: string
          format: date-time
          title: Fecha de alta
        parqueadero:
          allOf:
          - $ref: '#/components/schemas/AdminParqueaderoDTO'
          nullable: true
          readOnly: true
        documento:
          allOf:
          - $ref: '#/components/schemas/AdminDocumentoDTO'
          nullable: true
          readOnly: true
      required:
      - correo
      - documento
      - id
      - parqueadero
      - persona
    AdminCuentaPaginaDTO:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/AdminCuentaResumenDTO'
      required:
      - count
      - next
      - previous
      - results
    AdminCuentaResumenDTO:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        correo:
          type: string
          format: email
          maxLength: 254
        persona:
          $ref: '#/components/schemas/AdminPersonaDTO'
        is_active:
          type: boolean
          title: Activo
          description: Indica si el usuario debe ser tratado como activo. Desmarque
            esta opción en lugar de borrar la cuenta.
        correo_verificado:
          type: boolean
        onboarding_estado:
          $ref: '#/components/schemas/OnboardingEstadoEnum'
        parqueadero_nombre:
          type: string
          nullable: true
          readOnly: true
        date_joined:
          type: string
          format: date-time
          title: Fecha de alta
      required:
      - correo
      - id
      - parqueadero_nombre
      - persona
    AdminDocumentoAccesoDTO:
      type: object
      properties:
        url:
          type: string
          format: uri
      required:
      - url
    AdminDocumentoDTO:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre_original:
          type: string
          maxLength: 255
        mime_type:
          type: string
          maxLength: 100
        size_bytes:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        estado:
          $ref: '#/components/schemas/DocumentoEstadoEnum'
        motivo_rechazo:
          type: string
        uploaded_at:
          type: string
          format: date-time
          readOnly: true
        reviewed_at:
          type: string
          format: date-time
          nullable: true
        revisado_por:
          type: string
          nullable: true
          readOnly: true
      required:
      - id
      - mime_type
      - nombre_original
      - revisado_por
      - size_bytes
      - uploaded_at
    AdminParqueaderoDTO:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
          maxLength: 150
        descripcion:
          type: string
        habilitacion_estado:
          $ref: '#/components/schemas/HabilitacionEstadoEnum'
        estado_operativo:
          $ref: '#/components/schemas/EstadoOperativoEnum'
        motivo_rechazo:
          type: string
        calle_principal:
          type: string
        calle_secundaria:
          type: string
        numero_lote:
          type: string
        latitud:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
        longitud:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - calle_principal
      - calle_secundaria
      - id
      - latitud
      - longitud
      - nombre
      - numero_lote
      - updated_at
    AdminPersonaDTO:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 100
        apellido:
          type: string
          maxLength: 100
        tipo_identificacion:
          $ref: '#/components/schemas/TipoIdentificacionEnum'
        identificacion:
          type: string
          maxLength: 20
      required:
      - apellido
      - identificacion
      - nombre
      - tipo_identificacion
    AdminRechazoDTO:
      type: object
      properties:
        motivo:
          type: string
          maxLength: 1000
          minLength: 3
      required:
      - motivo
    AdminSolicitudDetalleDTO:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        correo:
          type: string
          format: email
          maxLength: 254
        persona:
          $ref: '#/components/schemas/AdminPersonaDTO'
        is_active:
          type: boolean
          title: Activo
          description: Indica si el usuario debe ser tratado como activo. Desmarque
            esta opción en lugar de borrar la cuenta.
        correo_verificado:
          type: boolean
        onboarding_estado:
          $ref: '#/components/schemas/OnboardingEstadoEnum'
        date_joined:
          type: string
          format: date-time
          title: Fecha de alta
        parqueadero:
          $ref: '#/components/schemas/AdminParqueaderoDTO'
        documento:
          $ref: '#/components/schemas/AdminDocumentoDTO'
      required:
      - correo
      - documento
      - id
      - parqueadero
      - persona
    AdminSolicitudPaginaDTO:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/AdminSolicitudResumenDTO'
      required:
      - count
      - next
      - previous
      - results
    AdminSolicitudResumenDTO:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        correo:
          type: string
          format: email
          maxLength: 254
        persona:
          $ref: '#/components/schemas/AdminPersonaDTO'
        is_active:
          type: boolean
          title: Activo
          description: Indica si el usuario debe ser tratado como activo. Desmarque
            esta opción en lugar de borrar la cuenta.
        onboarding_estado:
          $ref: '#/components/schemas/OnboardingEstadoEnum'
        parqueadero_nombre:
          type: string
        documento_estado:
          type: string
        actualizada_en:
          type: string
          format: date-time
      required:
      - actualizada_en
      - correo
      - documento_estado
      - id
      - parqueadero_nombre
      - persona
...
    DocumentoEstadoEnum:
      enum:
      - PENDIENTE
      - APROBADO
      - RECHAZADO
      type: string
      description: |-
        * `PENDIENTE` - Pendiente
        * `APROBADO` - Aprobado
        * `RECHAZADO` - Rechazado
...
    EstadoOperativoEnum:
      enum:
      - ABIERTO
      - LLENO
      - CERRADO
      - INACTIVO
      - FUERA_DE_SERVICIO
      type: string
      description: |-
        * `ABIERTO` - Abierto
        * `LLENO` - Lleno
        * `CERRADO` - Cerrado
        * `INACTIVO` - Inactivo
        * `FUERA_DE_SERVICIO` - Fuera de servicio
...
    HabilitacionEstadoEnum:
      enum:
      - BORRADOR
      - PENDIENTE
      - APROBADO
      - RECHAZADO
      type: string
      description: |-
        * `BORRADOR` - Borrador
        * `PENDIENTE` - Pendiente
        * `APROBADO` - Aprobado
        * `RECHAZADO` - Rechazado
...
    OnboardingEstadoEnum:
      enum:
      - CORREO_PENDIENTE
      - DATOS_INICIALES_PENDIENTES
      - REVISION_PENDIENTE
      - RECHAZADO
      - CONFIGURACION_PENDIENTE
      - ACTIVO
      - DESHABILITADO
      type: string
      description: |-
        * `CORREO_PENDIENTE` - Correo pendiente
        * `DATOS_INICIALES_PENDIENTES` - Datos iniciales pendientes
        * `REVISION_PENDIENTE` - Revision pendiente
        * `RECHAZADO` - Rechazado
        * `CONFIGURACION_PENDIENTE` - Configuracion pendiente
        * `ACTIVO` - Activo
        * `DESHABILITADO` - Deshabilitado
...
    TipoIdentificacionEnum:
      enum:
      - CI
      - PASAPORTE
      - RUC
      type: string
      description: |-
        * `CI` - Cedula
        * `PASAPORTE` - Pasaporte
        * `RUC` - RUC
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT